Make default crypto libraries optional
