+++ Cargo.toml
@@ -9,10 +9,10 @@ license = "MIT"
 keywords = ["lorawan", "iot", "lpwan", "parser", "lightweight"]
 
 [dependencies]
-aes = "0.3.2"
-cmac = "0.2.0"
+aes = { version = "0.3.2", optional = true }
+cmac = { version = "0.2.0", optional = true }
 generic-array = "0.12.0"
-heapless = "0.5.3"
+heapless = "0.5.4"
 
 [dev-dependencies]
 criterion = "0.3"
@@ -28,7 +28,9 @@ debug = true
 [features]
 default = ["full"]
 
-full = ["with-to-string", "with-downlink"]
+full = ["with-to-string", "with-downlink", "default-crypto"]
+
+default-crypto = ["aes", "cmac"]
 
 with-to-string = []
