+++ src/keys.rs
@@ -7,8 +7,6 @@
 // author: Ivaylo Petrov <ivajloip@gmail.com>
 
 use generic_array::{GenericArray, typenum::U16};
-use aes::Aes128;
-use aes::block_cipher_trait::BlockCipher;
 
 /// AES128 represents 128 bit AES key.
 #[derive(Debug, Default, PartialEq)]
@@ -63,55 +61,4 @@ pub trait CryptoFactory {
 
     /// Method that creates a MAC calculator.
     fn new_mac(&self, key: &AES128) -> Self::M;
-}
-
-pub type Cmac = cmac::Cmac<Aes128>;
-
-/// Provides a default implementation for build object for using the crypto functions.
-#[derive(Default,Debug, PartialEq)]
-pub struct DefaultFactory;
-
-impl CryptoFactory for DefaultFactory {
-    type E = Aes128;
-    type D = Aes128;
-    type M = Cmac;
-
-    fn new_enc(&self, key: &AES128) -> Self::E {
-        Aes128::new(GenericArray::from_slice(&key.0[..]))
-    }
-
-    fn new_dec(&self, key: &AES128) -> Self::D {
-        Aes128::new(GenericArray::from_slice(&key.0[..]))
-    }
-
-    fn new_mac(&self, key: &AES128) -> Self::M {
-        use cmac::Mac;
-        Cmac::new_varkey(&key.0[..]).unwrap()
-    }
-}
-
-impl Encrypter for Aes128 {
-    fn encrypt_block(&self, block: &mut GenericArray<u8, U16>) {
-        aes::block_cipher_trait::BlockCipher::encrypt_block(self, block);
-    }
-}
-
-impl Decrypter for Aes128 {
-    fn decrypt_block(&self, block: &mut GenericArray<u8, U16>) {
-        aes::block_cipher_trait::BlockCipher::decrypt_block(self, block);
-    }
-}
-
-impl Mac for Cmac {
-    fn input(&mut self, data: &[u8]) {
-        cmac::Mac::input(self, data);
-    }
-
-    fn reset(&mut self) {
-        cmac::Mac::reset(self);
-    }
-
-    fn result(self) -> GenericArray<u8, U16> {
-        cmac::Mac::result(self).code()
-    }
 }
