+++ tests/lorawan.rs
@@ -17,6 +17,7 @@ use lorawan::keys::*;
 use lorawan::maccommandcreator::*;
 use lorawan::maccommands::*;
 use lorawan::parser::*;
+use lorawan::default_crypto::DefaultFactory;
 
 fn phy_join_request_payload() -> Vec<u8> {
     let mut res = Vec::new();
